fix(SimpleHAL): resolve 37 audited bugs — Flash IRQ safety, WWDG/TIM guards, I2C/PWM/USART/ADC fixes
Fixes all confirmed issues from the SimpleHAL bug audit (see PLAN.md):

Critical (crash/hang/reset-loop risk):
- SimpleFlash: disable IRQ around every erase/program op (CH32V003 can't
  fetch instructions from Flash while busy — an ISR firing mid-op could
  fault); also unlock/lock once per multi-byte write instead of per byte,
  and dedupe the 3x duplicated page_buffer statics
- SimpleWWDG: clamp window <= counter to prevent a permanent reset loop;
  document that WWDG_Disable() can't truly disable the watchdog once armed
  (hardware limitation, not fixable in software)
- SimpleADC: lazy-init ADC_ReadVrefInt()/ADC_GetVDD() so they can't hang
  waiting on a clock that was never enabled
- SimpleTIM: complete the PWM/Timer ownership guard (claim in
  TIM_SimpleInit, check in TIM_AttachInterrupt/AdvancedInit/SetFrequency,
  release in TIM_DetachInterrupt) so the two modules can no longer collide
  on the same timer silently

Important:
- SimpleI2C: NULL/zero-length checks on Write/Read/RegMulti/Scan, detect
  NACK inside the wait loop instead of after a full timeout, add
  I2C_TryReadReg() for unambiguous error reporting, fix a PARTIAL_REMAP
  comment that contradicted the header and add a compile-time guard so
  SOP-8 can't select an unsupported pin remap
- SimpleI2C_Soft: bounded timeout on SCL release (bit-bang bus could hang
  forever on a stuck line)
- SimplePWM: clamp every channel's CCR when frequency changes (prevented
  stuck-HIGH output on other channels sharing the timer), set
  timer_base_init in PWM_AdvancedInit, remove the unverified partial-remap
  pin table that silently broke PWM output on both old and new pins
- SimpleUSART: interrupt-driven RX ring buffer — hardware buffer was 1
  byte, bytes were lost if the app didn't poll fast enough
- SimpleTIM_Ext: Time_t fields uint16_t -> uint32_t (RAW mode overflowed
  at ~18h)
- SimpleArduino: fixed dtostrf() double-rounding and precision>=9
  overflow; added opt-in random()/random(min,max) macros
- SimpleDMA: DMA_Reset now clears all 4 flags (was GL only, causing false
  IRQs on next use); guard against the Circular+MemToMem combo the
  hardware doesn't support

Minor: SimplePWR clears PDDS before WFI/WFE (prevents sleep silently
becoming standby), SimpleSPI/IWDG/PWR doc fixes for behavior that can't
be changed at the wrapper level (SPI speed comment vs actual PCLK2,
selective RCC flag clear not supported by the SDK).

Verified each finding against current source before fixing — 3 items
from the original audit (SysTick CTLR reserved-bit claim, DMA priority
comment, 1-Wire recovery delay placement) turned out to be false
positives on closer inspection and were left unchanged with rationale
documented in PLAN.md.

Build verified clean (0 errors) across all 109 source files after every
change.

+++ User/SimpleHAL/SimpleTIM_Ext.h
@@ -82,9 +82,9 @@ typedef enum {
  * @brief Time structure
  */
 typedef struct {
-    uint16_t hours;      /**< ชั่วโมง (0-1193 สำหรับ max time) */
-    uint16_t minutes;    /**< นาที (0-59 normalized, 0-71582788 raw) */
-    uint16_t seconds;    /**< วินาที (0-59 normalized, 0-4294967295 raw) */
+    uint32_t hours;      /**< ชั่วโมง (0-59 normalized, ไม่จำกัดค่า raw) */
+    uint32_t minutes;    /**< นาที (0-59 normalized, ไม่จำกัดค่า raw) */
+    uint32_t seconds;    /**< วินาที (0-59 normalized, ไม่จำกัดค่า raw) */
 } Time_t;
 
 /* ========== Buffer Size Recommendations ========== */
